Fix empty state
